More Architecture work done. Still needs more work to be completed and some dependencies need to be resolved. Organized files to further build up the core architecture of the app.
