+++ app/build.gradle
@@ -1,6 +1,7 @@
 plugins {
     id 'com.android.application'
     id 'org.jetbrains.kotlin.android'
+    id 'com.google.devtools.ksp' version "1.8.0-1.0.9"
 }
 
 android {
@@ -48,12 +49,23 @@ android {
 
 dependencies {
 
-    implementation 'androidx.core:core-ktx:1.7.0'
-    implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.3.1'
-    implementation 'androidx.activity:activity-compose:1.3.1'
-    implementation "androidx.compose.ui:ui:$compose_ui_version"
-    implementation "androidx.compose.ui:ui-tooling-preview:$compose_ui_version"
-    implementation 'androidx.compose.material:material:1.2.0'
+    //Import the Compose BOM
+    implementation platform('androidx.compose:compose-bom:2023.01.00')
+
+    implementation "androidx.core:core-ktx:1.9.0"
+
+    implementation "androidx.activity:activity-compose:1.6.1"
+    implementation "androidx.compose.ui:ui"
+    implementation "androidx.compose.ui:ui-tooling:1.3.3"
+    implementation "androidx.compose.ui:ui-tooling-preview:"
+    implementation "androidx.compose.material:material:1.3.1"
+
+    implementation "androidx.lifecycle.lifecycle:lifecycle-runtime-ktx:2.5.1"
+    implementation "androidx.lifecycle:lifecycle-viewmodel-compose:2.6.0"
+    implementation "androidx.navigation:navigation-compose:2.5.3"
+
+
+
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
@@ -63,6 +75,6 @@ dependencies {
 
     //Room
     implementation "androidx.room:room-runtime:$room_version"
-    //ksp "androidx.room:room-compiler:$room_version"
+    ksp "androidx.room:room-compiler:$room_version"
     implementation "androidx.room:room-ktx:$room_version"
 }
