+++ app/src/main/java/com/example/passtickmain/ui/AppViewModelProvider.kt
@@ -1,4 +1,4 @@
-package com.example.passtickmain.ui.theme
+package com.example.passtickmain.ui
 
 import android.app.Application
 import androidx.lifecycle.ViewModelProvider.AndroidViewModelFactory
@@ -6,17 +6,14 @@ import androidx.lifecycle.createSavedStateHandle
 import androidx.lifecycle.viewmodel.CreationExtras
 import androidx.lifecycle.viewmodel.initializer
 import androidx.lifecycle.viewmodel.viewModelFactory
-import com.example.passtickmain.InventoryApplication
-import com.example.passtickmain.ui.Account.HomeViewModel
-import com.example.passtickmain.ui.Account.ItemDetailsViewModel
-import com.example.passtickmain.ui.Account.AccountEditViewModel
-import com.example.passtickmain.ui.Account.AccountEntryViewModel
+
 
 /**
- * Provides Factory to create instance of ViewModel for the entire Inventory app
+ * Provides Factory to create instance of ViewModel for the entire PassStick app
  */
 object AppViewModelProvider {
     val Factory = viewModelFactory {
+
         // Initializer for ItemEditViewModel
         initializer {
             ItemEditViewModel(
@@ -44,7 +41,7 @@ object AppViewModelProvider {
 
 /**
  * Extension function to queries for [Application] object and returns an instance of
- * [InventoryApplication].
+ * [PasstickApplication].
  */
-fun CreationExtras.inventoryApplication(): InventoryApplication =
-    (this[AndroidViewModelFactory.APPLICATION_KEY] as InventoryApplication)
+fun CreationExtras.passStickApplication(): PassStickApplication =
+    (this[AndroidViewModelFactory.APPLICATION_KEY] as PassStickApplication)
