+++ build.gradle
@@ -1,6 +1,6 @@
 buildscript {
     ext {
-        compose_ui_version = '1.2.0'
+        compose_ui_version = '1.4.0'
         room_version = '2.5.0'
     }
 }// Top-level build file where you can add configuration options common to all sub-projects/modules.
@@ -8,4 +8,8 @@ plugins {
     id 'com.android.application' version '7.3.1' apply false
     id 'com.android.library' version '7.3.1' apply false
     id 'org.jetbrains.kotlin.android' version '1.7.0' apply false
+}
+
+task clean(type: Delete) {
+    delete rootProject.buildDir
 }
